+++ features/content/feed_component/src/main/java/com/tokopedia/feedcomponent/domain/usecase/GetProfileHeaderUseCase.kt
@@ -33,9 +33,9 @@ class GetProfileHeaderUseCase @Inject constructor(@ApplicationContext val contex
     companion object {
         const val PARAM_USER_ID_TARGET = "userIDTarget"
 
-        fun createRequestParams(targetUserId: Long): RequestParams {
+        fun createRequestParams(targetUserId: Int): RequestParams {
             val requestParams = RequestParams.create()
-            requestParams.putLong(PARAM_USER_ID_TARGET, targetUserId)
+            requestParams.putInt(PARAM_USER_ID_TARGET, targetUserId)
             return requestParams
         }
     }
